Make pipeline layer/device-portable, add feature labels, fix two run bugs
Parameterize experiment_0 by layer and make it run on both MPS (local) and
CUDA (server), add autointerp feature labels to every output, and fix two bugs
that only surfaced on the first real GPU run.

Layer + device portability
- config.py: LAYER from EXP0_LAYER env (default 6) derives the SAE id/hook,
  Neuronpedia + S3 URLs (SAE_SOURCE), and a per-layer outputs/layer_NN/ dir
  (RUN_DIR). EXP0_DEVICE override + is_mps() so mps / mps:0 / cuda:N classify
  correctly. cache_stats records the layer in saved metadata.
- run_metrics.py / qualitative_check.py / visualize.py write to RUN_DIR.
- .gitattributes: recursive LFS globs so outputs/layer_NN/*.pt|*.html track.
- migrate existing layer-6 artifacts into outputs/layer_06/.

Feature labels
- fetch_labels.py: bulk-download the layer's autointerp descriptions from the
  Neuronpedia S3 export into outputs/layer_NN/feature_labels.json.
- config.load_feature_labels() / feature_label() with "feature <idx>" fallback.
- run_metrics, visualize, qualitative_check render labels (superparents, top
  edges, sankey nodes); qualitative_check prefers the bulk file, API as fallback.

Bug fixes (found by the first CUDA run)
- cache_stats.py: cast recon error to acc_dtype - on CUDA the accumulators are
  float64 while the model is float32, so `fc.T @ err` hit a Double-vs-Float
  error. Hidden on MPS (both float32), so the code had only ever run on MPS.
- fetch_labels.py: discover the batch count from S3 instead of hardcoding 256 -
  the count varies per layer (layer 12 has 513), which under-fetched layer 12 to
  ~50% coverage. Now lists batch keys with pagination.

+++ cache_stats.py
@@ -154,7 +154,10 @@ def main():
         feats = sae.encode(resid)                           # [n, D_SAE]
         x_hat = sae.decode(feats)                           # [n, d_model]
         resid_err = resid - x_hat                           # [n, d_model]
-        err = (resid_err * resid_err).sum(dim=1)            # [n] base recon error
+        # Cast to acc_dtype: on CUDA/CPU the accumulators are float64 while the
+        # model runs in float32, so `fc.T @ err` below would hit a Double-vs-Float
+        # mismatch. (On MPS both are float32, which is why this only bit on CUDA.)
+        err = (resid_err * resid_err).sum(dim=1).to(acc_dtype)  # [n] base recon error
 
         fired = (feats > C.FIRE_THRESHOLD).to(acc_dtype)    # [n, D_SAE]
         g = per_token_ablation_gain(feats, resid_err, W_dec).to(acc_dtype)  # [n, D_SAE]
